# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AffiTrack - Affiliate Business Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /revenues:
    get:
      operationId: listRevenues
      tags: [revenues]
      summary: List revenues
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueList"
    post:
      operationId: createRevenue
      tags: [revenues]
      summary: Create a revenue entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
  /revenues/{id}:
    get:
      operationId: getRevenue
      tags: [revenues]
      summary: Get a revenue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
    patch:
      operationId: updateRevenue
      tags: [revenues]
      summary: Update a revenue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
    delete:
      operationId: deleteRevenue
      tags: [revenues]
      summary: Delete a revenue entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ADS
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseList"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # EMPLOYEES
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeList"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SCHEDULES
...
components:
  schemas:
...
    Revenue:
      type: object
      required: [id, date, amount, product, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        amount:
          type: number
        actualAmount:
          type: ["number", "null"]
        product:
          type: string
        units:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    RevenueInput:
      type: object
      required: [date, amount, product]
      properties:
        date:
          type: string
        amount:
          type: number
        actualAmount:
          type: ["number", "null"]
        product:
          type: string
        units:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    RevenueUpdate:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
        actualAmount:
          type: ["number", "null"]
        product:
          type: string
        units:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    RevenueList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Revenue"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ADS
...
    Expense:
      type: object
      required: [id, category, amount, date, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, amount, date]
      properties:
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
    ExpenseUpdate:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
    ExpenseList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # EMPLOYEES
    Employee:
      type: object
      required: [id, name, role, salaryType, salaryAmount, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        salaryType:
          type: string
          enum: [daily, monthly]
        salaryAmount:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [name, role, salaryType, salaryAmount]
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        salaryType:
          type: string
          enum: [daily, monthly]
        salaryAmount:
          type: number
        isActive:
          type: boolean
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        salaryType:
          type: string
          enum: [daily, monthly]
        salaryAmount:
          type: number
        isActive:
          type: boolean
    EmployeeList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # SCHEDULES
...