# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AffiTrack - Affiliate Business Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  category?: string;
  amount?: number;
  date?: string;
  /** @nullable */
  notes?: string | null;
}